+++ Gym management application/Class/QuanLyNhanVien/NhanVien.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Gym_management_appication.Class
+{
+    public class NhanVien
+    {
+        public int ID { get; set; }
+        public string HoTen { get; set; }
+        public string GioiTinh { get; set; }
+        public string Email { get; set; }
+
+        public long SoDT { get; set; }
+        public string DiaChi { get; set; }
+        public string ChucVu { get; set; }
+        public long Luong { get; set; }
+
+    }
+}
